refactor(logic): split combat_resolver into part-of concern files (Refs #3290)
Phase 0 prerequisite for the colonizethis_logic package split (#3290):
decompose combat/combat_resolver.dart (533 lines, > the 500-line Phase 0
target) into behavior-preserving `part`/`part of` concern files, mirroring
the established pattern (game_setup_create #3314, locked_province_assigner
#3316, capital_and_gp_fall in this PR).

- combat_resolver.dart (root, 228): imports/exports, the resolveBattleContext
  orchestration entry, and the three part directives.
- combat_resolver_engagement_outcome.dart (part): per-engagement outcome state
  transition (_applyEngagementOutcomeState) and mutual-annihilation garrison
  recovery (_recoverDefenderGarrisonIfNeeded).
- combat_resolver_post_battle.dart (part): post-battle assembly
  (_provinceOwnerIdAtBattleStart, _buildResolvedBattleGame,
  _buildPostBattleRegion).
- combat_resolver_support.dart (part): attacker initiative sort, the private
  _AttackingSideInBattle holder, and combat-math helpers
  (_incrementGeneralMedals, _defenderEffectiveLevel, _deploymentLimitForFaction,
  moraleMultiplierForGeneralMedals).

Pure structural extraction (no algorithm or public API change); all moved
members keep their original (public/private) visibility within the single
library via part-of, so external importers and tests are unaffected. Removed
the file from tool/logic_source_file_size_baseline.json so the 500-line gate
now actively enforces it (all four resulting files are <= 228 lines).

Refs #3290
