+++ packages/colonizethis_logic/lib/src/combat/combat_resolver.dart
@@ -21,6 +21,10 @@ export 'combat_constants.dart';
 export 'combat_engagement.dart' show resolveEngagement;
 export 'combat_types.dart';
 
+part 'combat_resolver_engagement_outcome.dart';
+part 'combat_resolver_post_battle.dart';
+part 'combat_resolver_support.dart';
+
 /// Resolves one battle context and returns updated Game state.
 /// SPEC/program/combat-resolution.md.
 Game resolveBattleContext(
@@ -221,313 +225,4 @@ Game resolveBattleContext(
   );
 
   return resolved;
-}
-
-({
-  String defenderFactionId,
-  String provinceOwnerId,
-  String? survivingAttackerFactionId,
-  String? currentDefenderGeneralId,
-  int currentDefenderMedals,
-  List<String> defenderUnitIds,
-})
-_applyEngagementOutcomeState({
-  required EngagementResult outcomeResult,
-  required _AttackingSideInBattle attacker,
-  required String? currentDefenderGeneralId,
-  required int currentDefenderMedals,
-  required String provinceOwnerId,
-  required String defenderFactionId,
-  required String? survivingAttackerFactionId,
-  required List<String> defenderUnitIds,
-  required int attackerIndex,
-  required List<_AttackingSideInBattle> attackerSidesWithMedals,
-  required int initialDefenderCount,
-  required int defenderEffectiveLevel,
-  required Map<String, General> generalsById,
-  required Map<String, Unit> unitsById,
-  required BattleContext ctx,
-}) {
-  switch (outcomeResult) {
-    case EngagementResult.attackerVictory:
-      _incrementGeneralMedals(
-        generalsById: generalsById,
-        generalId: attacker.assignedGeneralId,
-      );
-      final medalsAfterWin =
-          attacker.side.generalMedals +
-          (attacker.assignedGeneralId != null
-              ? kGeneralMedalsGainedOnBattleWin
-              : 0);
-      return (
-        defenderFactionId: attacker.side.factionId,
-        provinceOwnerId: attacker.side.factionId,
-        survivingAttackerFactionId: attacker.side.factionId,
-        currentDefenderGeneralId: attacker.assignedGeneralId,
-        currentDefenderMedals: medalsAfterWin > kGeneralMedalsMax
-            ? kGeneralMedalsMax
-            : medalsAfterWin,
-        defenderUnitIds: defenderUnitIds,
-      );
-    case EngagementResult.defenderVictory:
-      _incrementGeneralMedals(
-        generalsById: generalsById,
-        generalId: currentDefenderGeneralId,
-      );
-      var updatedMedals = currentDefenderMedals;
-      if (currentDefenderGeneralId != null) {
-        final updatedGeneral = generalsById[currentDefenderGeneralId];
-        if (updatedGeneral != null) {
-          updatedMedals = updatedGeneral.medals;
-        }
-      }
-      return (
-        defenderFactionId: defenderFactionId,
-        provinceOwnerId: provinceOwnerId,
-        survivingAttackerFactionId: null,
-        currentDefenderGeneralId: currentDefenderGeneralId,
-        currentDefenderMedals: updatedMedals,
-        defenderUnitIds: defenderUnitIds,
-      );
-    case EngagementResult.stalemate:
-      return (
-        defenderFactionId: defenderFactionId,
-        provinceOwnerId: provinceOwnerId,
-        survivingAttackerFactionId: survivingAttackerFactionId,
-        currentDefenderGeneralId: currentDefenderGeneralId,
-        currentDefenderMedals: currentDefenderMedals,
-        defenderUnitIds: defenderUnitIds,
-      );
-    case EngagementResult.mutualAnnihilation:
-      final updatedDefenderIds = _recoverDefenderGarrisonIfNeeded(
-        attackerIndex: attackerIndex,
-        attackerSidesWithMedals: attackerSidesWithMedals,
-        initialDefenderCount: initialDefenderCount,
-        defenderEffectiveLevel: defenderEffectiveLevel,
-        defenderUnitIds: defenderUnitIds,
-        provinceOwnerId: provinceOwnerId,
-        unitsById: unitsById,
-        provinceId: ctx.provinceId,
-      );
-      return (
-        defenderFactionId: provinceOwnerId,
-        provinceOwnerId: provinceOwnerId,
-        survivingAttackerFactionId: null,
-        currentDefenderGeneralId: currentDefenderGeneralId,
-        currentDefenderMedals: currentDefenderMedals,
-        defenderUnitIds: updatedDefenderIds,
-      );
-  }
-}
-
-List<String> _recoverDefenderGarrisonIfNeeded({
-  required int attackerIndex,
-  required List<_AttackingSideInBattle> attackerSidesWithMedals,
-  required int initialDefenderCount,
-  required int defenderEffectiveLevel,
-  required List<String> defenderUnitIds,
-  required String provinceOwnerId,
-  required Map<String, Unit> unitsById,
-  required String provinceId,
-}) {
-  final remainingAttackers = attackerSidesWithMedals.skip(attackerIndex + 1);
-  if (remainingAttackers.isEmpty) {
-    return defenderUnitIds;
-  }
-  final recoverCount = (initialDefenderCount * kGarrisonRecoveryFraction)
-      .ceil()
-      .clamp(1, initialDefenderCount);
-  final recoveryType = garrisonRecoveryRegimentTypeForEra(
-    defenderEffectiveLevel,
-  );
-  final updatedDefenderIds = List<String>.from(defenderUnitIds);
-  for (
-    var i = 0;
-    i < recoverCount && updatedDefenderIds.length < recoverCount;
-    i++
-  ) {
-    final id = '${kRecoveryUnitPrefix}${provinceId}_$i';
-    updatedDefenderIds.add(id);
-    unitsById[id] = Unit(
-      id: id,
-      type: recoveryType,
-      ownerId: provinceOwnerId,
-      locationProvinceId: provinceId,
-      medals: 0,
-    );
-  }
-  return updatedDefenderIds;
-}
-
-String? _provinceOwnerIdAtBattleStart(Game game, BattleContext ctx) {
-  final row = resolveProvinceRowForOwnershipTransfer(
-    game.worldState,
-    ctx.provinceId,
-  );
-  return row?.province.ownerId;
-}
-
-Game _buildResolvedBattleGame({
-  required Game game,
-  required BattleContext ctx,
-  required ({RegionData region, bool provinceChangedOwner}) post,
-  required String? survivingAttackerFactionId,
-  required Map<String, General> generalsById,
-  required CombatPhaseGeneralLedger ledger,
-  required String provinceOwnerAtBattleStart,
-}) {
-  var newWorldState = game.worldState.updateRegionById(
-    ctx.regionId,
-    (_) => post.region,
-  );
-
-  recordAttackCommandersForResolvedBattle(ctx, null, ledger);
-
-  var result = game.copyWith(
-    worldState: newWorldState,
-    generals: game.generals
-        .map((g) => generalsById[g.id] ?? g)
-        .toList(growable: false),
-  );
-  if (post.provinceChangedOwner && survivingAttackerFactionId != null) {
-    result = applyCanonicalSingleProvinceOwnershipTransfer(
-      result,
-      targetProvinceId: ctx.provinceId,
-      oldOwnerId: provinceOwnerAtBattleStart,
-      newOwnerId: survivingAttackerFactionId,
-    );
-  } else {
-    result = result.copyWith(
-      worldState: reconcileArmiesAfterUnitsChanged(result.worldState, result),
-    );
-  }
-  return result;
-}
-
-/// Builds post-battle region state: applies casualties, garrison recovery,
-/// province ownership change, and civilian cleanup when province changes hands.
-({RegionData region, bool provinceChangedOwner}) _buildPostBattleRegion({
-  required RegionData region,
-  required BattleContext ctx,
-  required Set<String> allCasualties,
-  required Map<String, Unit> unitsById,
-  required String defenderFactionId,
-  required String? survivingAttackerFactionId,
-  required List<String> defenderUnitIds,
-}) {
-  final survivingUnits = region.units
-      .where((u) => !allCasualties.contains(u.id))
-      .toList();
-
-  var updatedProvinces = region.provinces;
-  var provinceChangedOwner = false;
-  if (defenderUnitIds.isEmpty && survivingAttackerFactionId != null) {
-    provinceChangedOwner = provinceListContainsProvinceId(
-      updatedProvinces,
-      ctx.provinceId,
-    );
-  }
-
-  final recoveredUnits = unitsById.values
-      .where(
-        (u) =>
-            u.id.startsWith(kRecoveryUnitPrefix) &&
-            !allCasualties.contains(u.id),
-      )
-      .toList();
-  var finalUnits = [...survivingUnits, ...recoveredUnits];
-
-  final newRegion = RegionData(provinces: updatedProvinces, units: finalUnits);
-  return (region: newRegion, provinceChangedOwner: provinceChangedOwner);
-}
-
-void _sortAttackersByInitiative(
-  List<_AttackingSideInBattle> attackers,
-  Map<String, Unit> unitsById,
-  Random rng,
-) {
-  int cavalryCount(AttackingSide side) {
-    var count = 0;
-    for (final id in side.unitIds) {
-      final u = unitsById[id];
-      if (u == null) continue;
-      final stats = regimentStatsById(u.type);
-      if (stats != null && stats.isCavalry) count++;
-    }
-    return count;
-  }
-
-  final tieBreakRoll = rng.nextInt(kInitiativeTieBreakRngUpperExclusive);
-  int tieRank(String factionId) => Object.hash(factionId, tieBreakRoll);
-  final initiativeByAttacker = <_AttackingSideInBattle, double>{};
-  for (final attacker in attackers) {
-    final totalUnits = attacker.side.unitIds.length;
-    final cavalryShare = totalUnits > 0
-        ? cavalryCount(attacker.side) / totalUnits
-        : kZeroCavalryShareWhenNoUnits;
-    initiativeByAttacker[attacker] =
-        cavalryShare * initiativeCavalryShareWeight +
-        attacker.side.generalMedals * initiativeGeneralMedalWeight;
-  }
-
-  attackers.sort((a, b) {
-    final initA = initiativeByAttacker[a]!;
-    final initB = initiativeByAttacker[b]!;
-    final cmp = initB.compareTo(initA);
-    if (cmp != 0) return cmp;
-    return tieRank(a.side.factionId).compareTo(tieRank(b.side.factionId));
-  });
-}
-
-class _AttackingSideInBattle {
-  const _AttackingSideInBattle({
-    required this.side,
-    required this.assignedGeneralId,
-  });
-
-  final AttackingSide side;
-  final String? assignedGeneralId;
-}
-
-void _incrementGeneralMedals({
-  required Map<String, General> generalsById,
-  required String? generalId,
-}) {
-  if (generalId == null) return;
-  final current = generalsById[generalId];
-  if (current == null) return;
-  generalsById[generalId] = current.copyWith(
-    medals: (current.medals + kGeneralMedalsGainedOnBattleWin).clamp(
-      kGeneralMedalsMin,
-      kGeneralMedalsMax,
-    ),
-  );
-}
-
-int _defenderEffectiveLevel(Game game, String defenderFactionId) {
-  for (final m in game.minorNations) {
-    if (m.id == defenderFactionId) return m.effectiveMilitaryLevel;
-  }
-  for (final t in game.tribes) {
-    if (t.id == defenderFactionId) return t.effectiveMilitaryLevel;
-  }
-  return kDefaultEffectiveMilitaryEra;
-}
-
-/// Max regiments that can participate per side in one engagement.
-/// SPEC/game/military-generals.md: base 10; Nationalism tech → 12; +1 per general medal.
-int _deploymentLimitForFaction(Game game, String factionId, int generalMedals) {
-  final baseLimit =
-      game.playerById(factionId)?.techUnlocked?[kTechIdNationalism] == true
-      ? deploymentLimitWithNationalism
-      : deploymentLimitBase;
-  return baseLimit + generalMedals;
-}
-
-/// Morale aura from general medals. SPEC/game/military-generals.md:
-/// 5% per general medal, up to 20% (at 4 medals).
-double moraleMultiplierForGeneralMedals(int generalMedals) {
-  final capped = generalMedals.clamp(kGeneralMedalsMin, kGeneralMedalsMax);
-  return kMoraleMultiplierBaseFromGenerals +
-      (capped * kMoraleMultiplierPerGeneralMedal);
 }
