+++ tool/logic_source_file_size_baseline.json
@@ -1,6 +1,5 @@
 [
   "packages/colonizethis_logic/lib/src/ai/full_ai_civilian_work_selection_feedstock.dart",
-  "packages/colonizethis_logic/lib/src/combat/combat_resolver.dart",
   "packages/colonizethis_logic/lib/src/economy/world_market/deal_matcher.dart",
   "packages/colonizethis_logic/lib/src/orders/incremental_candidate_validator.dart",
   "packages/colonizethis_logic/lib/src/orders/order_suggestion_move_army.dart",
